moves spotpy scripts to specific directory
in order to make the package cleaner, I separated what was pute SMARTpy
scripts (staying in the root directory "scripts") and what was using
spotpy (in sripts/spotpy_). Formerly called SMARTpy_withSPOTPY is now
spot_lhs.py)

also needed to deal with the problem of running this script from
console/terminal, in this situation, and despite the presence of
__init__py, Python could not walk through the package anymore, because
spotpy is not in the same directory as SMARTpy anymore (only in a
subfolder). So now, when need to run scripts in terminal (e.g. to use
mpi), have to be in the directory root SMARTpy and use " set
PYTHONPATH=.  " in windows (ot replace "set" by "export" in linux) and
then run the scripts spotpy using a relative path from SMARTpy root
(e.g. mpiexec -n 4 python scripts/spotpy_/spotpy_lhs.py #Example 20 -p)

but because of this requirement to run scripts from SMARTpy root (so
that the imports of things from within this package in preambule can
work), it meant that path.realpath("../..") could not be used all the
time anymore. But when running spotpy from PyCharm, this instruction was
still needed => so moved definition of root_f in __main__ and then used
a if test to see if the current working directory CWD was equal to the
directory where spotpy scripts are, if so realpath(../..) still used,
otherwise, don't change the current dir and define current dir as the
root (assuming the user knows where to run spotpy from when using
console)