adds special case in rescale_irreg_cum_data
there was a special case that was not covered byt the previous algorithm. It happens when the discharge data provided as input starts on the date where simulation starts as well. It meant that the 'safety measure' of starting 2 days before start_report in get_dict_discharge_series had no effect because it would send a dataset still starting on start_report (because there was no data prior this date in input file). The way the algorithm is designed in  rescale_time_resolution_of_irregular_mean_data is to start at dt+1 and check the gap between dt0 and dt+1, which was creating no value for dt0 in the new_dict

I fixed this issue by adding an artificial date before start in my_dts series so that even for the first datetime, there is a backwards duplication of the value. I tested it and it yields the same results as before from dt+1 onwards, and also now I have the first value for the first datetime in the file.