includes NaN instead of empty string when no data
the algorithm that reads and rescales the observation data (discharge) used to add an empty string when data is missing (in increase_data_reso) and then decrease_data_reso was trying to find date/convert string to float, which meant that if data was missing the datetime was not included in the output dict.

Decided to change this and have the datetime in the dict with NaN as the value for the datetime keys where there is no data/not enough data to (decresase_data_reso). This will be easier to use in the future in objectivefunctions and also with spotpy.

Nota. because increase_data_reso now adds NaN and not an empty string in the dictionary, it means that the except TypeError in decrease_data_reso because useless because Python just returns a NaN when trying to do an mathematical operation involving a NaN value. But I decided to keep it anyway, to be extra safe, or in case I use the function from a different place or something.

Tested on #Example and it yields results as expected. Replaces empty values by NaN in output file, and also the dictionary now contains the complete timeseries of datetime in simulation period.