adds use of SMARTc if available
Adds the functionality to use SMARTc.onestep if the module is importable (i.e. if it is installed on the python distribution used), otherwise, falls back on the pure Python implementation of one_step available in SMARTpy. If SMARTc is available, SMARTpy automatically uses it, the user does not have the choice to go for the run_one_step in SMARTpy (since it is slower and gives the same results).