changes check for netCDF4 import
in some cases, on High Performance Clusters for example, the imp.find_module was raising the ImportError even if doing "import netCDF4" was working. So using an alternative solution using Dataset not equal to None. Because imp was not useful for netCDF4 anymore, got rid of it for 'csv' too because this is a built-in package