gets rid of unlimited dimension for netcdf file
When using parallel=True with netCDF4, it does not create a NETCDF4 (which is the default) but it creates a NETCDF3_CLASSIC instead. And NETCDF3_CLASSIC does not accept for unlimited dimensions to be anything else than the first dimensions in a multi dimensional variable. So there were 2 solutions: invert 'NbSamples' and 'DateTime' in the creation of the variable 'Simulations', or to not use an unlimited dimension for DateTime. Since the length of DateTime is known from the start, it was not actually required to use an unlimited dimension here, so opted for the second solution.

Adds explicitly format='NETCDF4' (which should be the default) but it seems to be overwritten by netCDF4 Dataset anyway.