allows for no gauged area provided
In single simulation mode, the SMART model can be run without observed discharge data and just yield simulated discharge from rain/peva and parameter values. However, SMARTpy was expecting a value for g_area in the settings file. Rather than asking the user to duplicate the c_area, it does it automatically. When gauged area and drainage area are the same, it will also let the user the possibility to leave out the gauged area. SMARTpy will know it means they are the same (or assumed to be).