move netcdf compression at the end of run
in netCDF4 v1.4.2, Dataset with parallel=True only works with 'NETCDF4', and zlib=True raises a RuntimeError: NetCDF: Invalid argument. The documentation for v1.4.2 clearly specifies that compression does not work in parallel mode, but it used to work in v1.4.0 where 'NETCDF3_CLASSIC' files were created.

For SMARTpy to be compatible with the latest version of NetCDF4, it was required to change the compression strategy. The database file created in not compressed in the first place, and once the simulations are complete, a copy of the .nc file is created with compression this time, and the old file is then deleted.