converts most dictionaries to nd arrays
In order to simplify the code and to make it more ready to include a C++ extension that runs the whole simulation loop, completely redesigned the model structure module.

it impacts the inout module because it needs to work with nd arrays now, not dictionaries anymore. Also got rid of the 'report' and 'method' because they became irrelevant (reporting at the simu_gap can be achieved by the user by setting the save_time_gap to the simu_time_gap, and the method (summary/raw) can be decided using the keyword argument in SMART.simulate).

it also impacts on MonteCarlo module which already converted dict to numpy arrays, so it simplifies this module as well.