adds check on warmup duration against sim duration
SMARTpy is designed to use the first X days of the simulation period as made-up warm-up data to initialise the states of the model. Which implies that if the duration of the simulation is shorter than the duration of the warm-up period, the collected meteorological data for the simulation period will fall short to simulate the whole warm-up duration. This condition is now check before running the warm-up simulation to ensure than an IndexError out of bounds will not be raised, and provide the user with alternative work-around solutions.