corrects mistake on check existence of nd array
since release of v0.2.0 and the change to numpy arrays as the way to store the timeseries to accommodate for the communication with the new smartcpp (v0.2.0) for allsteps (commit e6969c0cf42c9136acad5437e397c84d630912a1), the additional test on the existence of the evaluation data was made on a numpy array that does not work like a list that evaluates as True/False if it is not empty/empty, so it was raising a ValueError. It is now fixed by testing more explicitly if nd_flow remained as None or was assigned a numpy array. Duplicated this for nd_discharge for the simulation array as well.