applies same check as for getters in file writer
For consistency, and also because the checks that were made where a bit convoluted compared to what check was needed, decided to harmonise checks on data availability with the previous commit in the write_output_files, i.e. testing the existence of the numpy arrays rather than those of the outputs tuple and the flow dictionary.

Took the opportunity to rephrase the Exception message to be more specific.