fixes bug by checking that start/end and saving timedelta are coherent
When working with unusual time deltas and or start/end datetimes, there may be situations where the specified end datetime is not a multiple of timedelta from the start datetime. In this situation, the reshape of nd_storage in the function run_all_steps of the module structure.py.

This is now fixed by adding a check method in the TimeFrame class. If the condition above is not met, this will raise an Exception and inform the user of the latest datetime that should be used in place of the one given.